# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DreamCar Reserva de vehiculos",
        "description": "Esta API nos permitira ver y manipular informacion a las reservas de los vehiculos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/marcas/show/{id}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Obtener una marcas por id",
                "description": "Renorna los detalles de una marca",
                "operationId": "4b10c8c6988d876ad72642e9a18ef6ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la marca con id seleccionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "marca": {
                                            "type": "string",
                                            "maxLength": 50,
                                            "example": "Nombre de una Marca",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/marcas/marca/{marca}": {
            "get": {
                "tags": [
                    "Marcas"
                ],
                "summary": "Buscar marca por nombre",
                "description": "Retorna una marca",
                "operationId": "97688254c60ea360308e75e3dd90f1c0",
                "parameters": [
                    {
                        "name": "marca",
                        "in": "path",
                        "description": "Nombre de la marca a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "marca": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Nombre de una Marca",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la marca",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/reservas/show/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener una reserva por id",
                "description": "Renorna los detalles de una reserva",
                "operationId": "c73e3697aaef74220acea5b4394ccdf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la reserva con id seleccionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "email": {
                                            "type": "string",
                                            "maxLength": 50,
                                            "example": "Email de la reserva",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reservas/reserva/{email_cliente}": {
            "get": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Buscar reserva por email del cliente",
                "description": "Retorna una reserva",
                "operationId": "eb9632714b93b250adecd2a2abeccec3",
                "parameters": [
                    {
                        "name": "email_cliente",
                        "in": "path",
                        "description": "email de la reserva a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Email de la reserva",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reservas/reserva/{id}": {
            "get": {
                "tags": [
                    "Reserva"
                ],
                "summary": "Mostrar los detalles de una reserva",
                "description": "Retorna los detalles de una reserva",
                "operationId": "fa6c34db9e516106f87624fcce1414f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la reserva a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "type": "string",
                                                "maxLength": 1000,
                                                "example": "Detalles de la reserva",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la Reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}